+++ android/src/main/kotlin/com/destinyed/flutter_youtube_downloader/DownloadManagerClass.kt
@@ -1,36 +1,48 @@
-package com.destinyed.flutter_youtube_downloader
+package com.example.tubmate.Constants
 
 import android.app.DownloadManager
 import android.content.Context
+import android.content.Context.DOWNLOAD_SERVICE
 import android.net.Uri
 import android.os.Environment
+import android.widget.Toast
+import androidx.core.content.ContextCompat.getSystemService
 
-class DownloadManagerClass(ctx: Context) {
 
-    var manager = ctx.getSystemService(Context.DOWNLOAD_SERVICE) as DownloadManager
-    private var downloadID : Long? = null
+class DownloadManagerClass(val ctx: Context) {
 
-    fun download(nTitle: String, nDescription: String, nUrl: String){
+    var downloadManager = ctx.getSystemService(DOWNLOAD_SERVICE) as DownloadManager
+    private var downloadID: Long? = null
 
-        //get the manager
+    fun download(nTitle: String, nDescription: String, nUrl: String) {
 
-        var request = DownloadManager.Request((Uri.parse(nUrl))).apply {
+        try {
+            val request = DownloadManager.Request(Uri.parse(nUrl))
 
-            //set notification definition
-            setDescription(nDescription)
-            setTitle(nTitle)
+            //Setting title of request
+            request.setTitle(nTitle)
 
-            allowScanningByMediaScanner()
-            setNotificationVisibility(DownloadManager.Request.VISIBILITY_VISIBLE_NOTIFY_COMPLETED)
+            //Setting description of request
+            request.setDescription("Your file is downloading")
 
-            setVisibleInDownloadsUi(true)
-            setDestinationInExternalPublicDir(Environment.DIRECTORY_DOWNLOADS, "$nTitle.mp4")
+            //set notification when download completed
+            request.setNotificationVisibility(DownloadManager.Request.VISIBILITY_VISIBLE_NOTIFY_COMPLETED)
+
+            //Set the local destination for the downloaded file to a path within the application's external files directory
+//            request.setDestinationInExternalPublicDir(Environment.DIRECTORY_DOWNLOADS, "$nTitle.mp4")
+            request.setDestinationInExternalFilesDir(ctx, Environment.DIRECTORY_DOWNLOADS, nTitle + "mp4")
+            request.allowScanningByMediaScanner()
+
+
+            //Enqueue download and save the referenceId
+            downloadID = downloadManager.enqueue(request)
+
+            Toast.makeText(ctx, "Video is downloading... Slide down to see progress", Toast.LENGTH_LONG).show()
+
+        } catch (e: IllegalArgumentException) {
+            Toast.makeText(ctx, "Download link is broken or not availale for download", Toast.LENGTH_LONG).show()
+
         }
-
-        downloadID = manager.enqueue(request)
-
-
-
 
 
     }