+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ## 0.0.1
 
 * A simple easy to use Flutter Youtube link extractor and downloader
+
+## O.O.3
+
+* Added instruction for extrating and downloading video on Android API 29+